Add tests for services

+++ src/main/java/ua/foodtracker/service/MealService.java
@@ -9,7 +9,7 @@ import java.util.Optional;
 public interface MealService {
     Page<Meal> findAllByPage(String pageNumber);
 
-    long pageCount();
+    int pageCount();
 
     void add(Meal meal);
